Replicate secret on startup of the config-updater
This way the registry secret that is initially configured in eirini's
configmap will be replicated in the app namespace so that it is
available before it is updated for the first time.

If the secret is not there the config-updater logs a warning but does
not fail. This way if someone updates the value in the configmap, the
updater will replicate the secret.

[#173013466]